+++ src/Model/Entity/PhotosAlbum.php
@@ -41,18 +41,28 @@ class PhotosAlbum extends Entity
      * Virtual fields that should be exposed
      * @var array
      */
-    protected $_virtual = ['path'];
+    protected $_virtual = ['path', 'preview'];
 
     /**
      * Gets the album full path (virtual field)
-     * @return string|void
+     * @return string
      */
     protected function _getPath()
     {
-        if (empty($this->_properties['id'])) {
+        return PHOTOS . $this->_properties['id'];
+    }
+
+    /**
+     * Gets the album preview (virtual field)
+     * @return string|void
+     * @since 2.21.1
+     */
+    protected function _getPreview()
+    {
+        if (empty($this->_properties['photos'])) {
             return;
         }
 
-        return PHOTOS . $this->_properties['id'];
+        return collection($this->_properties['photos'])->extract('path')->first();
     }
 }