+++ src/View/View/AppView.php
@@ -88,8 +88,8 @@ class AppView extends View
         parent::initialize();
 
         //Loads helpers
-        $this->loadHelper(METOOLS . '.BBCode');
-        $this->loadHelper(METOOLS . '.Breadcrumbs');
+        $this->loadHelper(ME_TOOLS . '.BBCode');
+        $this->loadHelper(ME_TOOLS . '.Breadcrumbs');
         $this->loadHelper(RECAPTCHA_MAILHIDE . '.Mailhide');
         $this->loadHelper(ME_CMS . '.Widget');
     }
@@ -108,6 +108,8 @@ class AppView extends View
      */
     public function renderLayout($content, $layout = null)
     {
+        $this->plugin = ME_CMS;
+
         $this->setBlocks();
 
         //Assign the userbar