+++ tests/TestCase/Controller/AppControllerTest.php
@@ -73,9 +73,9 @@ class AppControllerTest extends IntegrationTestCase
         $this->assertEquals([
             'Cake\Controller\Component\CookieComponent',
             ME_CMS . '\Controller\Component\AuthComponent',
-            METOOLS . '\Controller\Component\FlashComponent',
+            ME_TOOLS . '\Controller\Component\FlashComponent',
             'Cake\Controller\Component\RequestHandlerComponent',
-            METOOLS . '\Controller\Component\UploaderComponent',
+            ME_TOOLS . '\Controller\Component\UploaderComponent',
             'Recaptcha\Controller\Component\RecaptchaComponent',
         ], $components);