+++ tests/TestCase/View/Cell/PhotosWidgetsCellTest.php
@@ -68,7 +68,7 @@ class PhotosWidgetsCellTest extends TestCase
 
         $result = $this->Widget->widget($widget)->render();
         $expected = [
-            ['div' => ['class' => 'widget']],
+            ['div' => ['class' => 'widget mb-4']],
             'h4' => ['class' => 'widget-title'],
             'Albums',
             '/h4',
@@ -95,7 +95,7 @@ class PhotosWidgetsCellTest extends TestCase
         //Renders as list
         $result = $this->Widget->widget($widget, ['render' => 'list'])->render();
         $expected = [
-            ['div' => ['class' => 'widget']],
+            ['div' => ['class' => 'widget mb-4']],
             'h4' => ['class' => 'widget-title'],
             'Albums',
             '/h4',
@@ -161,13 +161,13 @@ class PhotosWidgetsCellTest extends TestCase
 
         $result = $this->Widget->widget($widget)->render();
         $expected = [
-            ['div' => ['class' => 'widget']],
+            ['div' => ['class' => 'widget mb-4']],
             'h4' => ['class' => 'widget-title'],
             'Latest photo',
             '/h4',
             ['div' => ['class' => 'widget-content']],
-            'a' => ['href' => '/albums', 'class' => 'thumbnail', 'title' => ''],
-            'img' => ['src', 'alt', 'class' => 'img-responsive'],
+            'a' => ['href' => '/albums'],
+            'img' => ['src', 'alt', 'class' => 'img-fluid thumbnail'],
             '/a',
             '/div',
             '/div',
@@ -177,16 +177,16 @@ class PhotosWidgetsCellTest extends TestCase
         //Tries another limit
         $result = $this->Widget->widget($widget, ['limit' => 2])->render();
         $expected = [
-            ['div' => ['class' => 'widget']],
+            ['div' => ['class' => 'widget mb-4']],
             'h4' => ['class' => 'widget-title'],
             'Latest 2 photos',
             '/h4',
             ['div' => ['class' => 'widget-content']],
-            ['a' => ['href' => '/albums', 'class' => 'thumbnail', 'title' => '']],
-            ['img' => ['src', 'alt', 'class' => 'img-responsive']],
+            ['a' => ['href' => '/albums']],
+            ['img' => ['src', 'alt', 'class' => 'img-fluid thumbnail']],
             '/a',
-            ['a' => ['href' => '/albums', 'class' => 'thumbnail', 'title' => '']],
-            ['img' => ['src', 'alt', 'class' => 'img-responsive']],
+            ['a' => ['href' => '/albums']],
+            ['img' => ['src', 'alt', 'class' => 'img-fluid thumbnail']],
             '/a',
             '/div',
             '/div',
@@ -229,13 +229,13 @@ class PhotosWidgetsCellTest extends TestCase
 
         $result = $this->Widget->widget($widget)->render();
         $expected = [
-            ['div' => ['class' => 'widget']],
+            ['div' => ['class' => 'widget mb-4']],
             'h4' => ['class' => 'widget-title'],
             'Random photo',
             '/h4',
             ['div' => ['class' => 'widget-content']],
             ['a' => ['href' => '/albums', 'class' => 'thumbnail', 'title' => '']],
-            ['img' => ['src', 'alt', 'class' => 'img-responsive']],
+            ['img' => ['src', 'alt', 'class' => 'img-fluid']],
             '/a',
             '/div',
             '/div',
@@ -245,16 +245,16 @@ class PhotosWidgetsCellTest extends TestCase
         //Tries another limit
         $result = $this->Widget->widget($widget, ['limit' => 2])->render();
         $expected = [
-            ['div' => ['class' => 'widget']],
+            ['div' => ['class' => 'widget mb-4']],
             'h4' => ['class' => 'widget-title'],
             'Random 2 photos',
             '/h4',
             ['div' => ['class' => 'widget-content']],
             ['a' => ['href' => '/albums', 'class' => 'thumbnail', 'title' => '']],
-            ['img' => ['src', 'alt', 'class' => 'img-responsive']],
+            ['img' => ['src', 'alt', 'class' => 'img-fluid']],
             '/a',
             ['a' => ['href' => '/albums', 'class' => 'thumbnail', 'title' => '']],
-            ['img' => ['src', 'alt', 'class' => 'img-responsive']],
+            ['img' => ['src', 'alt', 'class' => 'img-fluid']],
             '/a',
             '/div',
             '/div',